+++ python_bindings/src/bindings.cpp
@@ -32,6 +32,7 @@ void bind_robot_model(nb::module_ &m);
 void bind_tasks(nb::module_ &m);
 void bind_kinematics_solver(nb::module_ &m);
 void bind_pose_metrics(nb::module_ &m);
+void bind_acceleration_solver(nb::module_ &m);
 
 NB_MODULE(_embodik_impl, m) {
   m.doc() = "embodiK: High-performance inverse kinematics with Pinocchio";
@@ -840,4 +841,7 @@ NB_MODULE(_embodik_impl, m) {
 
   // Bind pose metrics free functions
   bind_pose_metrics(m);
+
+  // Bind acceleration-level solver
+  bind_acceleration_solver(m);
 }